Add platform package
